EES-5091 Abstract host builder configuration to ProcessorHostBuilder extension method in preparation for adding integration tests.
